910739 - let system.scheduleReboot API return action id

+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/system/test/SystemHandlerTest.java
@@ -1854,7 +1854,7 @@ public class SystemHandlerTest extends BaseHandlerTestCase {
 
         int preScheduleSize = dr.size();
 
-        Integer returnInt = handler.scheduleReboot(adminKey,
+        Long returnInt = handler.scheduleReboot(adminKey,
                 new Integer(testServer.getId().intValue()), new Date());
         assertEquals(returnInt, new Integer(1));
